Face Recognition
encode_faces.py needs to run on a dataset (collection of images) and create encoding.pickle. Once we have obtained encoding .pickle we will run recognize_faces_images on a example image file which we can store in example folder. This will return name of the people in the image.